LinGui: fix additional mingw gtk UI build issues

+++ gtk/src/main.c
@@ -616,7 +616,7 @@ IoRedirect(signal_user_data_t *ud)
     // Open activity log.
     config = ghb_get_user_config_dir(NULL);
     pid = getpid();
-    path = g_strdup_printf("%s/Activity.log.%d", config, pid);
+    path = g_strdup_printf("%s/Activity.log.%"PRId64, config, (long)pid);
     ud->activity_log = g_io_channel_new_file (path, "w", NULL);
     ud->job_activity_log = NULL;
     str = g_strdup_printf("<big><b>%s</b></big>", path);
